+++ src/kikuchipy/__init__.py
@@ -1,4 +1,5 @@
-# Copyright 2019-2024 The kikuchipy developers
+#
+# Copyright 2019-2025 the kikuchipy developers
 #
 # This file is part of kikuchipy.
 #
@@ -14,6 +15,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
+#
 
 import lazy_loader
 
@@ -22,6 +24,7 @@ credits = [
     "Håkon Wiik Ånes",
     "Lars Andreas Hastad Lervik",
     "Ole Natlandsmyr",
+    "Thomas G. Woodcock",
     "Tina Bergh",
     "Eric Prestat",
     "Andreas V. Bugten",
@@ -30,7 +33,7 @@ credits = [
     "Carter Francis",
     "Magnus Nord",
 ]
-__version__ = "0.11.2"
+__version__ = "0.12.dev2"
 
 __getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)