+++ src/kikuchipy/constants.py
@@ -1,4 +1,5 @@
-# Copyright 2019-2024 The kikuchipy developers
+#
+# Copyright 2019-2025 the kikuchipy developers
 #
 # This file is part of kikuchipy.
 #
@@ -14,6 +15,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
+#
 
 """Constants and such useful across modules."""
 
@@ -54,7 +56,7 @@ try:  # pragma: no cover
         pyopencl_context_available = False
     else:
         pyopencl_context_available = True
-except:  # pragma: no cover
+except Exception:  # pragma: no cover
     # Have to use bare except because PyOpenCL might raise its own
     # LogicError, but we also want to catch import errors here
     pyopencl_context_available = False
@@ -67,6 +69,6 @@ try:
     from numpy.exceptions import VisibleDeprecationWarning
 except ImportError:  # pragma: no cover
     # Removed in NumPy 2.0.0
-    from numpy import VisibleDeprecationWarning
+    from numpy import VisibleDeprecationWarning  # noqa: F401
 
 del dep_version, deps_for_version_check, version